+++ test/endpoint.js
@@ -76,6 +76,7 @@ describe('phonetics', function() {
   require('./phonetics/metaphone.js');
   require('./phonetics/mra.js');
   require('./phonetics/nysiis.js');
+  require('./phonetics/onca.js');
   require('./phonetics/phonex.js');
   require('./phonetics/roger-root.js');
   require('./phonetics/soundex.js');